+++ src/Axle/Services/INotificationService.cs
@@ -4,12 +4,14 @@
 namespace Axle.Services
 {
     using System;
+    using Axle.Contracts;
+    using Axle.Dto;
 
     public interface INotificationService
     {
 #pragma warning disable CA1710 // Event name should end in EventHandler
-        event Action<int> OnSessionTerminated;
+        event Action<TerminateSessionNotification> OnSessionTerminated;
 
-        void PublishSessionTermination(int sessionId);
+        void PublishSessionTermination(TerminateSessionNotification terminateSessionNotification);
     }
 }